Move GagGlobals to core
