+++ Gag.py
@@ -20,9 +20,10 @@ from typing import List, Optional
 from core.AttackGlobals import GROUP
 from core.Exceptions import (GagCountError, LockedGagError,
                              LockedGagTrackError, NotEnoughGagsError)
-from GagGlobals import (DEFAULT_GAG_COUNT, DEFAULT_TRACK_EXPS_CURRENT, GAG,
-                        GAG_CARRY_LIMITS, GAG_DAMAGE, GAG_LABELS,
-                        GAG_TRACK_LABELS, LEVELS, MULTI_TARGET_GAGS, TRACK)
+from core.GagGlobals import (DEFAULT_GAG_COUNT, DEFAULT_TRACK_EXPS_CURRENT,
+                             GAG, GAG_CARRY_LIMITS, GAG_DAMAGE, GAG_LABELS,
+                             GAG_TRACK_LABELS, LEVELS, MULTI_TARGET_GAGS,
+                             TRACK)
 
 
 def get_default_gag_count() -> List[List[int]]:
