+++ core/GagGlobals.py
